backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.10",
    "langchain-huggingface>=1.2.1",
    "langgraph>=1.0.10",
    "langgraph-cli[inmem]>=0.4.14",
    "tavily-python>=0.7.22",
    "transformers[torch]>=5.3.0",
]
